Full Client/Server interaction is available! But problems persist
Currently, the UI is drawn from the metadata and not actual data. This
is a slight oversight that can be sorta fixed - depends on how we want
to solve this particular problem. One way is to mirror the
ExperimentManager object between client and server, the second, to keep
a simple memory object that remembers which experiment the client has
selected/is a part of. I favor the latter, but maybe the former is more
robust?

Also, I added comments.