Consent Text Wall updates based on consent. Packet also sent to server.
Changelog:
	- ServerEnforcer receives a packet from GuiConsent that has the consent
status
	- For now, the JSON data doesn't matter, but it could contain their
answers or other information about the conditions of the player's
consent (i.e. I authorize limited sharing, etc.)
	- TextWallRenderHandler displays text based on the status of the
player's consent.
	- Player, when logging in, is forced to read the GuiConsent if they are
joining an experiment server. in the future, use the REST API to
determine whether the player has already given consent. If so, they
should NOT be forced to re-read that and should instead simply see the
ExperimentGUI pop up. This is handled during renderTick on
ClientEnforcer. Is there a better place for that??