Add command result to API, tp_to command, Set gamerules
Command result is assumed to be successful until proven otherwise. This
logic will be inverted later with the planned command system overhaul
TP_TO command allows an agent to teleport to objects (blocks or
entities) in an experiment area. Used as an alternative to an agent
handling navigation.
Some gamerules are set to prevent annoyances/obsticles for agents. Such
as the day/night cycle being disabled, preventing weather, and mob
spawning.