Added pathways to Wall featurs
Pathways can be open, open (full height), door.
All pathway types can be reversed to be built in the opposite direction.
Example, a door on the west side of a wall is the reverse of a door on
the east side.  Door blocks can be defined as doorBlockName and
doorBlockMeta; however, doorBlockMeta is not used because typicaly the
meta is used in positioning the door.  There also are blockName and Meta
fields for use with buttons, pressure plates, levers, and wall blocks.